feat(review): implement MarkScoreSelector and SuggestionBuilder for improved suggestion generation
